requirements.txt
aioredis==1.3.1
amqp==2.5.2
apns2==0.5.0
asgiref==3.5.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
boto3==1.18.16
botocore==1.21.16
celery==4.4.2
certifi==2019.6.16
cffi==1.14.0
channels==3.0.5
channels-redis==3.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.8
daphne==3.0.2
decorator==4.4.2
Django==2.2
django-common-helpers==0.9.2
django-cors-headers==3.0.2
django-cron==0.5.1
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-push-notifications==2.0.0
django-rest-swagger==2.1.2
django-simple-history==2.7.2
django-sslserver==0.20
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
fcm-django==0.3.4
ffmpeg-python==0.2.0
flower==0.9.3
future==0.18.2
h2==2.6.2
hiredis==1.0.1
hpack==3.0.0
html2text==2020.1.16
html5lib==1.0.1
http-ece==1.1.0
httplib2==0.20.4
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==1.5.0
incremental==21.3.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
kombu==4.6.8
MarkupSafe==1.1.1
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.0.4
numpy==1.21.2
openapi-codec==1.3.2
opencv-python==4.6.0.66
pdfkit==0.6.1
Pillow==8.2.0
proglog==0.1.9
psycopg2-binary==2.8.4
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pyfcm==1.4.7
PyGetWindow==0.0.9
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.0
python-decouple==3.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2019.1
pywebpush==1.10.1
redis==3.4.1
reportlab==3.5.32
requests==2.22.0
s3transfer==0.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stripe==2.40.0
tornado==5.1.1
tqdm==4.60.0
Twisted==21.7.0
txaio==20.4.1
typing-extensions==3.10.0.0
tzlocal==2.0.0
uritemplate==3.0.0
vine==1.3.0
webencodings==0.5.1
websocket-client==0.57.0
xhtml2pdf==0.2.3
zipp==3.1.0
zope.interface==5.1.0
